Move AddConsoleCommandPass from FrameworkBundle to Console.
