+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -34,7 +34,7 @@
     "require-dev": {
         "symfony/asset": "~2.8|~3.0",
         "symfony/browser-kit": "~2.8|~3.0",
-        "symfony/console": "~2.8.8|~3.0.8|~3.1.2|~3.2",
+        "symfony/console": "~3.3",
         "symfony/css-selector": "~2.8|~3.0",
         "symfony/dom-crawler": "~2.8|~3.0",
         "symfony/polyfill-intl-icu": "~1.0",
@@ -56,7 +56,8 @@
     },
     "conflict": {
         "phpdocumentor/reflection-docblock": "<3.0",
-        "phpdocumentor/type-resolver": "<0.2.0"
+        "phpdocumentor/type-resolver": "<0.2.0",
+        "symfony/console": "<3.3"
     },
     "suggest": {
         "ext-apcu": "For best performance of the system caches",
