+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+3.3.0
+-----
+
+* added `AddConsoleCommandPass` (originally in FrameworkBundle)
+
 3.2.0
 ------
